Medicare-Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic[email]
cryptography
pycryptodome
requests
python-dotenv
ipfshttpclient
sqlalchemy
alembic
passlib[bcrypt]
uvicorn
